It just gets worse.
Who in their right mind decided that blocks are a good idea. Turns out,
you don't include the <style> or <script> tags?? Therefore, there's zero
syntax highlighting.

You are punished for being modular. I'm sure there's a way to sort this,
but jesus is this framework bad. I'll be sticking with Next.js!